[Refactor/#155] refact oauth2FailureHandler
