+++ src/main/java/com/example/waggle/domain/member/repository/MemberRepository.java
@@ -1,6 +1,7 @@
 package com.example.waggle.domain.member.repository;
 
 import com.example.waggle.domain.member.entity.Member;
+import com.example.waggle.global.security.oauth2.OAuth2UserInfoFactory.AuthProvider;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
@@ -11,6 +12,8 @@ public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByUsername(String username);
 
     Optional<Member> findByEmail(String email);
+
+    Optional<Member> findByEmailAndAuthProvider(String email, AuthProvider authProvider);
 
     Optional<Member> findByNickname(String nickname);
